Utils::strtol only used in the app
