+++ src/utils.cpp
@@ -98,15 +98,4 @@ namespace Util {
         return b.substr(idx);
     }
 
-    bool strtol(const char* nptr, long& n)
-    {
-        if (!nptr || *nptr == '\0') return false;
-        char* endptr = nullptr;
-        long tmp = std::strtol(nptr, &endptr, 10);
-        if (*endptr != '\0') return false;
-        if (tmp == LONG_MAX || tmp == LONG_MIN) return false;
-        n = tmp;
-        return true;
-    }
-
 }                                       // namespace Util
