+++ src/utils.hpp
@@ -63,13 +63,6 @@ namespace Util {
              to the end of the string.
      */
     std::string suffix(const std::string& path);
-
-    /*!
-      @brief Convert a C string to a long value, which is returned in n.
-             Returns true if the conversion is successful, else false.
-             n is not modified if the conversion is unsuccessful. See strtol(2).
-     */
-    bool strtol(const char* nptr, long& n);
 }                                       // namespace Util
 
 #endif                                  // #ifndef UTILS_HPP_
